+++ CHANGELOG.md
@@ -5,9 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2021-11-20
+
+### Added
+
+- Support for multiple paths
+- Improved error handling
+- Allow for specific snapshot to be restored
+- Docker image
+
+### Fixed
+
+- rclone in docker volumes
+
+### Changed
+
+- [Breaking Change] Declaration of docker volumes. See: https://autorestic.vercel.app/migration/1.4_1.5
+- [Breaking Change] Hooks default executing directory now defaults to the config file directory. See: https://autorestic.vercel.app/migration/1.4_1.5
+
 ## [1.4.1] - 2021-10-31
 
-### Fixes
+### Fixed
 
 - Numeric values from config files not being passed to env.